Default param to empty object in call if undefined
The param parameter will only be defaulted to null in broker.call if it is passed in as undefined instead of null or undefined. This will preserve backwards compatibility for users who were not passing in params and expecting it to be defaulted but also allow a caller to set it to null.

To go along with this change the types on the Context class were made stricter. Param is no longer a union type with null but you must pass in a type since the default is now object instead of GenericObject which can really be anything. A default was left in place to avoid needing to make cascading changes to the Context types but the default of object makes the context class almost useless without providing a type.