Allow service broker to be reclaimed by GC
Changes to add tracing and to collect garbage collection metrics were
causing the ServiceBroker instances to not be reclaimed after going out
of scope.

The use of AsyncStorage in both the service broker and the Tracer class
were causing AsyncStorage to hold the ServiceBroker. The broker's
start/stop lifecycle hooks were modified to enable the scope and to
disable the scope and clear the store on stop.

The automatic import and invocation of the gc-stats module in the metrics module
was causing the ServiceBroker to be held by system/Context in our unit
tests. I am not sure of the exact reason but only importing and invoking
the module when metrics are enabled resolved this issue